requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
twilio==8.10.0
lxml>=5.0.0
playwright>=1.45.0
gunicorn==21.2.0
Werkzeug==3.0.1
